Update setup.sh
add more check

+++ setup.sh
@@ -66,11 +66,15 @@ echo '[INSTALL] Installing Jadx'
 gh-release-install \
 'skylot/jadx' \
 'jadx-{version}.zip' \
-'./mobsf/StaticAnalyzer/tools/jadx.zip' \
---version-file './mobsf/StaticAnalyzer/tools/jadx-{version}.ver'
-unzip -qq -d ./mobsf/StaticAnalyzer/tools/jadx/ ./mobsf/StaticAnalyzer/tools/jadx.zip
-rm -f ./mobsf/StaticAnalyzer/tools/jadx.zip
-
+'./mobsf/StaticAnalyzer/tools/jadx.zip'
+if [ -f './mobsf/StaticAnalyzer/tools/jadx.zip' ]; then
+    rm -fr ./mobsf/StaticAnalyzer/tools/jadx/
+    unzip -qq -d ./mobsf/StaticAnalyzer/tools/jadx/ ./mobsf/StaticAnalyzer/tools/jadx.zip
+    rm -f ./mobsf/StaticAnalyzer/tools/jadx.zip
+else
+    echo '[ERROR] Problem dowloading Jadx'
+    exit 1
+fi
 
 echo '[INSTALL] Clean Up'
 bash scripts/clean.sh y
